refactor: introduce the withFadeAnimation() HOC
